requirements.txt
nltk
pyenchant